added

+++ C++/C++(With_Phitron)/Module_03/class_2.cpp
@@ -6,7 +6,7 @@ class Student
 public:
     char name[100];
     int roll;
-    float gpa;
+    double gpa;
 };
 
 int main()
